# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CreateRestaurantRequest:
    properties:
      address:
        type: string
      cuisineType:
        type: string
      deliveryFee:
        type: number
      description:
        type: string
      email:
        type: string
      image:
        type: string
      maxDeliveryTime:
        type: integer
      minDeliveryTime:
        type: integer
      name:
        type: string
      phone:
        type: string
      priceRange:
        maximum: 3
        minimum: 1
        type: integer
    required:
    - address
    - cuisineType
    - email
    - name
    - phone
    - priceRange
    type: object
...
  handlers.UpdateRestaurantRequest:
    properties:
      address:
        type: string
      cuisineType:
        type: string
      deliveryFee:
        type: number
      description:
        type: string
      email:
        type: string
      image:
        type: string
      isOpen:
        type: boolean
      maxDeliveryTime:
        type: integer
      minDeliveryTime:
        type: integer
      name:
        type: string
      phone:
        type: string
      priceRange:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive food delivery application API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Restaurant App API
  version: "1.0"
paths:
...
  /public/restaurants:
    get:
      consumes:
      - application/json
      description: Get all active restaurants with pagination and filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by cuisine type
        in: query
        name: cuisine
        type: string
      - description: Search in restaurant names
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all restaurants
      tags:
      - restaurants
  /public/restaurants/{id}:
    get:
      consumes:
      - application/json
      description: Get a single restaurant with its details
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get restaurant by ID
      tags:
      - restaurants
...
  /restaurants:
    post:
      consumes:
      - application/json
      description: Create a new restaurant (restaurant owners only)
      parameters:
      - description: Restaurant data
        in: body
        name: restaurant
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateRestaurantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Create a new restaurant
      tags:
      - restaurants
  /restaurants/{id}:
    put:
      consumes:
      - application/json
      description: Update restaurant details (owner only)
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: string
      - description: Restaurant update data
        in: body
        name: restaurant
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateRestaurantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Update restaurant
      tags:
      - restaurants
...
  /restaurants/me:
    get:
      consumes:
      - application/json
      description: Get restaurant owned by current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get current user's restaurant
      tags:
      - restaurants
  /restaurants/search:
    get:
      consumes:
      - application/json
      description: Search restaurants with various filters like cuisine, price range,
        rating, etc.
      parameters:
      - description: Search query (name or cuisine)
        in: query
        name: q
        type: string
      - description: Cuisine type filter
        in: query
        name: cuisine
        type: string
      - description: Minimum rating (0-5)
        in: query
        name: minRating
        type: number
      - description: Maximum price range (1-4)
        in: query
        name: maxPrice
        type: number
      - description: Maximum delivery fee
        in: query
        name: deliveryFee
        type: number
      - description: Filter by open status
        in: query
        name: isOpen
        type: boolean
      - description: 'Sort by: rating, delivery_fee, delivery_time'
        enum:
        - rating
        - delivery_fee
        - delivery_time
        in: query
        name: sortBy
        type: string
      - description: 'Sort order: asc, desc'
        enum:
        - asc
        - desc
        in: query
        name: sortOrder
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Search restaurants with advanced filters
      tags:
      - restaurants
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"